backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
alembic==1.12.1
psycopg[binary]==3.1.18
asyncpg==0.29.0
qdrant-client==1.7.0
openai==1.3.5
pydantic==2.5.0
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
bcrypt==5.0.0
python-multipart==0.0.6
redis==5.0.1
python-dotenv==1.0.0
httpx==0.25.2
email-validator==2.3.0
PyPDF2==3.0.1
pdfplumber==0.11.9
pdfminer.six==20251230
pypdfium2>=4.18.0
caldav==1.4.0
boto3==1.34.0
pandas>=2.0.0
openpyxl>=3.1.0
xlrd>=2.0.0
Pillow>=10.0.0